Midterm concept
I want to create a hairband with a LED matrix on it which can show emotions. Using the infrared sensor to change the state of LED matrix to show different emotions. At first I wanted to use LCD to show the words of emotions like "happy", "sad". But I tried a lot of times and the LCD didn't work. So I decided to use LED matrix to show simple emotions, like "smiling face","sad face", which is also smaller and easier to be attached to the hairband.